+++ packages/amis-core/src/utils/style-helper.ts
@@ -361,3 +361,61 @@ export function insertEditCustomStyle(customStyle: any, id?: string) {
     'wrapperCustomStyle-' + (id?.replace('u:', '') || uuid())
   );
 }
+
+export interface InsertCustomStyle {
+  themeCss?: any;
+  classNames?: CustomStyleClassName[];
+  id?: string;
+  defaultData?: any;
+  customStyleClassPrefix?: string;
+}
+
+export class StyleDom {
+  id: string;
+  constructor(id: string) {
+    this.id = id;
+  }
+  /**
+   * 插入自定义样式
+   *
+   * @param {InsertCustomStyle} params - 插入自定义样式的参数
+   * @param {string} params.themeCss - 主题样式
+   * @param {string} params.classNames - 自定义样式类名
+   * @param {string} params.defaultData - 默认数据
+   * @param {string} params.customStyleClassPrefix - 自定义样式类名前缀
+   */
+  insertCustomStyle({
+    themeCss,
+    classNames,
+    defaultData,
+    customStyleClassPrefix
+  }: InsertCustomStyle) {
+    insertCustomStyle(
+      themeCss,
+      classNames,
+      this.id,
+      defaultData,
+      customStyleClassPrefix
+    );
+  }
+
+  /**
+   * 插入外层自定义样式
+   *
+   * @param wrapperCustomStyle 自定义样式
+   */
+  insertEditCustomStyle(wrapperCustomStyle: any) {
+    insertEditCustomStyle(wrapperCustomStyle, this.id);
+  }
+  /**
+   * 移除自定义样式
+   */
+  removeCustomStyle(type?: string) {
+    const style = document.getElementById(
+      (type ? type + '-' : '') + this.id.replace('u:', '')
+    );
+    if (style) {
+      style.remove();
+    }
+  }
+}
