+++ packages/amis-ui/scss/components/_page.scss
@@ -17,7 +17,6 @@
   &-content {
     padding: var(--Page-content-paddingY) var(--Page-content-paddingX);
     position: relative;
-    height: 100%;
     width: 100%;
   }
